refactor(ffi): make GraphQlEvent non-lossy and move it to graphql module
- Make the GraphQL Event -> GraphQlEvent conversion fallible (TryFrom) so a
  malformed bcs payload surfaces an error through events() instead of silently
  decoding to empty contents; this makes the 'non-lossy' doc accurate.
- Move GraphQlEvent into graphql::query_types alongside the other GraphQL
  query-result types (Epoch, Validator, EventFilter, ...).
- Clarify the doc: explain why package_id/module/sender are null (system/genesis,
  0x0 reported as null) vs timestamp (not yet in a checkpoint), and that
  type_/contents/data/json are always present.
